+++ bash/feeders/adsbexchange.sh
@@ -34,8 +34,10 @@
 ## VARIABLES
 
 BUILDDIR=$PWD
+MLATCLIENTDIR="$BUILDDIR/mlat-client"
 ADSBEXCHANGEDIR="$BUILDDIR/adsbexchange"
 
+source ../bash/variables.sh
 source ../bash/functions.sh
 
 ## INFORMATIVE MESSAAGE ABOUT THIS SOFTWARE
@@ -44,58 +46,155 @@ clear
 
 echo -e "\033[31m"
 echo "-----------------------------------------------------"
-echo " Now ready to set up ADS-B Exchange feed."
+echo " Now ready to set up a feed to ADS-B Exchange."
 echo "-----------------------------------------------------"
 echo -e "\033[33mADSBexchange.com is a co-op of ADS-B/Mode S/MLAT feeders from around the world."
-echo "PiAware is required to be installed in order to feed this site. If PiAware is not"
-echo "currently installed this script will execute the PiAware installation script after"
-echo "which this script will continue with the ADS-B Exchange feed setup."
 echo ""
 echo "http://www.adsbexchange.com/how-to-feed/"
-echo "https://github.com/flightaware/piaware"
 echo -e "\033[37m"
 read -p "Press enter to continue..." CONTINUE
 
 ## CHECK FOR PREREQUISITE PACKAGES
 
 echo -e "\033[33m"
-echo "Installing packages needed to build and fulfill dependencies..."
+echo "Installing packages needed to fulfill dependencies..."
 echo -e "\033[37m"
+CheckPackage curl
+CheckPackage build-essential
+CheckPackage debhelper
+CheckPackage python3-dev
 CheckPackage netcat
 
-## CONFIGURE PIAWARE TO FEED ADS-B EXCHANGE IF PIAWARE IS INSTALLED
+## DOWNLOAD OR UPDATE THE MLAT-CLIENT SOURCE
 
-echo -e "\e[33m"
-printf "Configuring PiAware if it is installed..."
-if [ $(dpkg-query -W -f='${Status}' piaware 2>/dev/null | grep -c "ok installed") != 0 ]; then
+# Check if the git repository already exists locally.
+if [ -d $MLATCLIENTDIR ] && [ -d $MLATCLIENTDIR/.git ]; then
+    # A directory with a git repository containing the source code exists.
     echo -e "\033[33m"
-    echo "Adding the ADS-B Exchange feed to PiAware's configuration..."
-    ORIGINALFORMAT=`sudo piaware-config -show | grep mlatResultsFormat | sed 's/mlatResultsFormat //g'`
-    MLATRESULTS=`sed 's/[{}]//g' <<< $ORIGINALFORMAT`
-    CLEANFORMAT=`sed 's/ beast,connect,feed.adsbexchange.com:30005//g' <<< $MLATRESULTS`
-    sudo piaware-config -mlatResultsFormat "${CLEANFORMAT} beast,connect,feed.adsbexchange.com:30005"
-    echo "Restarting PiAware so new configuration takes effect..."
+    echo "Updating the local mlat-client git repository..."
     echo -e "\033[37m"
-    sudo piaware-config -restart
-    echo ""
+    cd $MLATCLIENTDIR
+    git pull
+    git checkout tags/${MLATCLIENTTAG}
+else
+    # A directory containing the source code does not exist in the build directory.
+    echo -e "\033[33m"
+    echo "Cloning the mlat-client git repository locally..."
+    echo -e "\033[37m"
+    git clone https://github.com/mutability/mlat-client.git
+    cd $MLATCLIENTDIR
+    git checkout tags/${MLATCLIENTTAG}
 fi
 
-## CONFIGURE SCRIPT TO EXECUTE NETCAT TO FEED ADS-B EXCHANGE
+## BUILD AND INSTALL THE MLAT-CLIENT PACKAGE
 
-echo -e "\033[33mSetting permissions on adsbexchange-maint.sh..."
+echo -e "\033[33m"
+echo "Building the mlat-client package..."
 echo -e "\033[37m"
-sudo chmod +x $ADSBEXCHANGEDIR/adsbexchange-maint.sh
+dpkg-buildpackage -b -uc
 
-echo -e "\033[33mAdding startup line to rc.local..."
+echo -e "\033[33m"
+echo "Installing the mlat-client package..."
 echo -e "\033[37m"
-lnum=($(sed -n '/exit 0/=' /etc/rc.local))
-((lnum>0)) && sudo sed -i "${lnum[$((${#lnum[@]}-1))]}i ${ADSBEXCHANGEDIR}/adsbexchange-maint.sh &\n" /etc/rc.local
+sudo dpkg -i ${BUILDDIR}/mlat-client_${MLATCLIENTVERSION}*.deb
 
-## START NETCAT ADS-B EXCHANGE FEED
+## REMOVE THE OLD ADB-B EXCHANGE STARTUP LINE FROM /ETC/RC.LOCAL IF IT EXISTS
 
-echo -e "\033[33mExecuting adsbexchange-maint.sh..."
+# Check if the old adsbexchange-maint.sh line exists in /etc/rc.local.
+if grep -Fxq "${ADSBEXCHANGEDIR}/adsbexchange-maint.sh &" /etc/rc.local; then
+    # Kill any currently running instances of the adsbexchange_maint.sh script.
+    PIDS=`ps -efww | grep -w "${ADSBEXCHANGEDIR}/adsbexchange-maint.sh &" | awk -vpid=$$ '$2 != pid { print $2 }'`
+    if [ ! -z "$PIDS" ]; then
+        sudo kill $PIDS >> $LOGFILE
+        sudo kill -9 $PIDS >> $LOGFILE
+    fi
+    # Remove the old line from /etc/rc.local.
+    sudo sed -i /$$ADSBEXCHANGEDIR\/adsbexchange-maint.sh &/d /etc/rc.local
+fi
+
+## CONFIGURE SCRIPT TO EXECUTE AND MAINTAIN MLAT-CLIENT AND NETCAT TO FEED ADS-B EXCHANGE
+
+# Ask the user for the user name for this receiver.
+ADSBEXCHANGEUSER=$(whiptail --backtitle "$BACKTITLETEXT" --title "ADS-B Exchange User Name" --nocancel --inputbox "\nPlease enter your ADS-B Exchange user name.\n\nIf you have more than one receiver, this username should be unique.\nExample: \"username-01\", \"username-02\", etc." 12 78 3>&1 1>&2 2>&3)
+
+# Get the altitude of the receiver from the Google Maps API using the latitude and longitude assigned dump1090-mutability.
+RECEIVERLAT=`GetConfig "LAT" "/etc/default/dump1090-mutability"`
+RECEIVERLON=`GetConfig "LON" "/etc/default/dump1090-mutability"`
+
+# Ask the user for the receivers altitude. (This will be prepopulated by the altitude returned from the Google Maps API.
+RECEIVERALT=$(whiptail --backtitle "$BACKTITLETEXT" --title "Receiver Longitude" --nocancel --inputbox "\nEnter your recivers atitude." 9 78 "`curl -s https://maps.googleapis.com/maps/api/elevation/json?locations=$RECEIVERLAT,$RECEIVERLON | python -c "import json,sys;obj=json.load(sys.stdin);print obj['results'][0]['elevation'];"`" 3>&1 1>&2 2>&3)
+
+# Create the adsbexchange directory in the build directory if it does not exist.
+if [ ! -d "$ADSBEXCHANGEDIR" ]; then
+    mkdir $ADSBEXCHANGEDIR
+fi
+
+echo -e "\033[33mCreating the file adsbexchange-netcat_maint.sh..."
 echo -e "\033[37m"
-sudo $ADSBEXCHANGEDIR/adsbexchange-maint.sh &
+tee $ADSBEXCHANGEDIR/adsbexchange-netcat_maint.sh > /dev/null <<EOF
+#! /bin/sh
+while true
+  do
+    sleep 30
+    /bin/nc 127.0.0.1 30005 | /bin/nc feed.adsbexchange.com 30005
+  done
+EOF
+
+echo -e "\033[33mCreating the file adsbexchange-mlat_maint.sh..."
+echo -e "\033[37m"
+tee $ADSBEXCHANGEDIR/adsbexchange-mlat_maint.sh > /dev/null <<EOF
+#! /bin/sh
+while true
+  do
+    sleep 30
+    /usr/bin/mlat-client --input-type dump1090 --input-connect 127.0.0.1:30005 --lat $RECEIVERLAT --lon $RECEIVERLON --alt $RECEIVERALT --user $ADSBEXCHANGEUSER --server feed.adsbexchange.com:31090 --no-udp --results beast,connect,127.0.0.1:30104
+  done
+EOF
+
+echo -e "\033[33mSetting permissions on adsbexchange-netcat_maint.sh..."
+echo -e "\033[37m"
+sudo chmod +x $ADSBEXCHANGEDIR/adsbexchange-netcat_maint.sh
+
+echo -e "\033[33mSetting permissions on adsbexchange-mlat_maint.sh..."
+echo -e "\033[37m"
+sudo chmod +x $ADSBEXCHANGEDIR/adsbexchange-mlat_maint.sh
+
+echo -e "\033[33mAdding netcat startup line to rc.local..."
+echo -e "\033[37m"
+if ! grep -Fxq "$ADSBEXCHANGEDIR/adsbexchange-netcat_maint.sh &" /etc/rc.local; then
+    lnum=($(sed -n '/exit 0/=' /etc/rc.local))
+    ((lnum>0)) && sudo sed -i "${lnum[$((${#lnum[@]}-1))]}i ${ADSBEXCHANGEDIR}/adsbexchange-netcat_maint.sh &\n" /etc/rc.local
+fi
+echo -e "\033[33mAdding mlat-client startup line to rc.local..."
+echo -e "\033[37m"
+if ! grep -Fxq "$ADSBEXCHANGEDIR/adsbexchange-mlat_maint.sh &" /etc/rc.local; then
+    lnum=($(sed -n '/exit 0/=' /etc/rc.local))
+    ((lnum>0)) && sudo sed -i "${lnum[$((${#lnum[@]}-1))]}i ${ADSBEXCHANGEDIR}/adsbexchange-mlat_maint.sh &\n" /etc/rc.local
+fi
+
+## START THE MLAT-CLIENT AND NETCAT FEED
+
+# Kill any currently running instances of the adsbexchange-netcat_maint.sh script.
+PIDS=`ps -efww | grep -w "adsbexchange-netcat_maint.sh" | awk -vpid=$$ '$2 != pid { print $2 }'`
+if [ ! -z "$PIDS" ]; then
+    sudo kill $PIDS
+    sudo kill -9 $PIDS
+fi
+
+# Kill any currently running instances of the adsbexchange-mlat_maint.sh script.
+PIDS=`ps -efww | grep -w "adsbexchange-mlat_maint.sh" | awk -vpid=$$ '$2 != pid { print $2 }'`
+if [ ! -z "$PIDS" ]; then
+    sudo kill $PIDS
+    sudo kill -9 $PIDS
+fi
+
+echo -e "\033[33mExecuting adsbexchange-netcat_maint.sh..."
+echo -e "\033[37m"
+sudo $ADSBEXCHANGEDIR/adsbexchange-netcat_maint.sh > /dev/null &
+
+echo -e "\033[33mExecuting adsbexchange-mlat_maint.sh..."
+echo -e "\033[37m"
+sudo $ADSBEXCHANGEDIR/adsbexchange-mlat_maint.sh > /dev/null &
 
 echo -e "\033[33mConfiguration of the ADS-B Exchange feed is now complete."
 echo "Please look over the output generated to be sure no errors were encountered."