+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 The following is a history of the changes made to this project.
 
+## v2.2.0 *(August 31st, 2016)*
+
+* ADS-B Exchange script now sets up mlat-client to connect to their mlat-server.
+* Added the ability to disply either the dump1090-fa or dump1090-mutability map.
+* Changed the bash function which retrieves config file variables so it works properly.
+* Fixed issue causing a package installation error when PiAware versions change.
+
+## v2.1.0 *(August 30th, 2016)*
+
+* Added flight information side bar to plot map. *(advanced)*
+* Added flight data API to web site. *(advanced)*
+* Fixed an issue where settings were not being saved properly during upgrades.
+* The PiAware script has been modified to support the installation of PiAware 3.
+* PiAware is no longer automatically configured to share MLAT data with 3rd parties.
+
 ## v2.0.3 *(May 19th, 2016)*
 
 * Added password confirmation for MySQL database user to bash script. *(advanced)*